backend/requirements.txt
asgiref==3.2.10
Django==2.2.19
pytz==2021.3
sqlparse==0.4.2
django-extra-fields==3.0.2
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
drf-spectacular==0.18.0
drf-yasg==1.21.4
gunicorn==20.1.0
isort==5.9.3
itypes==1.2.0
Pillow==8.3.1
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.1.0
python-dotenv==0.19.2
requests==2.26.0
testfixtures==6.18.1
pyyaml ==6.0
